+++ SqExpress/SqlExport/Internal/PgSqlBuilder.cs
@@ -115,6 +115,12 @@ namespace SqExpress.SqlExport.Internal
             return this.VisitExprDerivedTableValuesCommon(derivedTableValues, parent);
         }
 
+        public override bool VisitExprInsert(ExprInsert exprInsert, IExpr? parent)
+        {
+            this.GenericInsert(exprInsert, null, null);
+            return true;
+        }
+
         public override bool VisitExprInsertOutput(ExprInsertOutput exprInsertOutput, IExpr? parent)
         {
             this.GenericInsert(exprInsertOutput.Insert,