+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.1.0
+### New Features
+- the DTO code-generator now has a parameter that allows generating C# records: ```--model-type ImmutableCalss|Record``` or ```<SqModelGenType>ImmuatbleClass|Record</SqModelGenType>```;
+- "CheckExistenceBy" in the Insert data builder what adds WHERE EXISTS(...) to a Insert source to avoid duplicates inserting;
+
 # 0.3.0.0
 ### New Features
 - "ReadAsString(recordReader:ISqDataRecordReader): string" and "FromString(value: string?): ExprLiteral" are added to TableColumn. They allow performing mass export/import of database data.